refactor(web): move agent-compose types to infra/agent-compose
